fix(tests): check that the int and uint setters actually set the value
The previous version of the tests wasn't actually setting a value inside
the param; if you wrapped `int_data` or `uint_data` in an `OSSLParam`
and tried to `get()` from it, you would get back `None`, not a `Some` of
the new value, because there was no memory backing the `data` field.
(To avoid depending on the correctness of the implementation of
`OSSLParam::get()`, these tests dereference the inner param struct's
`data` pointer directly instead, but the principle is the same.)

This commit also removes a trailing underscore from the name of the
`test_uint_data` test, and a comment that restated the logic of the old
version of the tests (which were already imo pretty readable).